move bench into subdir
